fix(context): revert structural snapshot rehydration and rely on baseline gc
