+++ packages/core/src/context/config/profiles.ts
@@ -4,7 +4,6 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import { createStateSnapshotHydrationProcessor } from '../processors/stateSnapshotHydrationProcessor.js';
 import type {
   AsyncPipelineDef,
   ContextManagementConfig,
@@ -89,29 +88,6 @@ export const generalistProfile: ContextProfile = {
   ): PipelineDef[] =>
     // Helper to merge default options with dynamically loaded processorOptions by ID
     [
-      {
-        name: 'Initialization Hydration',
-        triggers: ['initialization'],
-        processors: [
-          createStateSnapshotHydrationProcessor('StateSnapshotHydration', env, {
-            target: (() => {
-              const res = resolveProcessorOptions(config, 'StateSnapshotSync', {
-                target: 'max',
-                maxStateTokens: 4000,
-                maxSummaryTurns: 5,
-              }).target;
-              if (
-                res === 'incremental' ||
-                res === 'freeNTokens' ||
-                res === 'max'
-              ) {
-                return res;
-              }
-              return undefined;
-            })(),
-          }),
-        ],
-      },
       {
         name: 'Immediate Sanitization',
         triggers: ['new_message'],
